requirements.txt
flask 
flask-cors
transformers 
pillow 
torch